requirements.txt
psycopg2==2.7.1
Django==1.11.2
smarturls==0.1.7
mongoengine==0.13.0
djangorestframework==3.6.3
